backend/requirements.txt
anthropic
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
python-multipart
openpyxl
python-docx
